Allow the secret to be safely specified externally.
The way the code is written today expects users to modify the file
that's under the version control system. This is inherently dangerous as
a developer needs to locally edit this file and then be very careful not
to accidentally commit the change in to the repository.

A better way to do this is to either allow these constants to be defined
through a compiler switch (for interactive dekstop development) or in a
separate file (outside the ForcePad source tree and the -I option allows
you to have this import statement resolve the header file correctly.)

This will also work better in a build server, where an in-place
modification to a file is tricky.